Changes to auto refresh the images every 30s
